+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/frontend/pinbally/PinballYConnector.java
@@ -377,12 +377,6 @@ PlayfieldWindow.Minimized = 0
   }
 
   @Override
-  public Playlist savePlaylist(Playlist playlist) {
-    //TODO not implemented yet
-    return null;
-  }
-
-  @Override
   protected void savePlaylistGame(int gameId, Playlist pl) {
     PinballYStatisticsParser parser = new PinballYStatisticsParser(this);
     parser.writePlaylistGame(getGame(gameId), pl);